+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next Release
 
+- Fix an issue in Foundry v13 that was preventing some chat message cards from working properly ([#1060](https://github.com/ben/foundry-ironsworn/pull/1060))
+
 ## 1.25.4
 
 - Fix a styling issue with Foundry rolltable-result editors ([#1059](https://github.com/ben/foundry-ironsworn/pull/1059))